REFACTOR: 집안일 멤버별 날짜별 완료율 조회 API 도메인 변경
