+++ src/main/java/com/simleetag/homework/api/domain/home/member/MemberService.java
@@ -1,11 +1,19 @@
 package com.simleetag.homework.api.domain.home.member;
 
+import java.time.LocalDate;
+import java.time.Period;
+import java.util.ArrayList;
+import java.util.List;
+
 import com.simleetag.homework.api.common.exception.HomeJoinException;
 import com.simleetag.homework.api.domain.home.Home;
 import com.simleetag.homework.api.domain.home.member.dto.MemberModifyRequest;
 import com.simleetag.homework.api.domain.home.member.repository.MemberRepository;
 import com.simleetag.homework.api.domain.user.User;
 import com.simleetag.homework.api.domain.user.UserService;
+import com.simleetag.homework.api.domain.work.task.TaskDslRepository;
+import com.simleetag.homework.api.domain.work.task.TaskStatus;
+import com.simleetag.homework.api.domain.work.task.api.TaskRateResponse;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -22,6 +30,8 @@ public class MemberService {
     private final MemberRepository memberRepository;
 
     private final MemberFinder memberFinder;
+
+    private final TaskDslRepository taskDslRepository;
 
     public Member join(Home home, Long userId) {
         final User user = userService.findById(userId);
@@ -51,4 +61,21 @@ public class MemberService {
         member.expire();
         return member;
     }
+
+    public List<TaskRateResponse> calculateTaskRatesByDueDates(Long memberId, LocalDate startDate, LocalDate endDate) {
+        Period period = Period.between(startDate, endDate);
+        List<TaskRateResponse> list = new ArrayList<>();
+        for (int i = 0; i < period.getDays() + 1; i++) {
+            double allTasks = taskDslRepository.findAllWithTaskGroupByHomeIdAndOwnerAndDueDate(memberId, startDate.plusDays(i)).size();
+            double doneTasks = taskDslRepository.findAllWithTaskGroupByHomeIdAndOwnerAndDueDate(memberId, startDate.plusDays(i)).stream().filter(task -> task.getTaskStatus().equals(TaskStatus.COMPLETED)).count();
+            double rate = 0;
+            if (allTasks != 0) {
+                rate = doneTasks / allTasks * 100.0;
+            }
+            TaskRateResponse response = new TaskRateResponse(startDate.plusDays(i), (int) rate);
+            list.add(response);
+        }
+        return list;
+    }
+
 }
