+++ src/main/java/com/simleetag/homework/api/domain/work/CategoryService.java
@@ -1,7 +1,6 @@
 package com.simleetag.homework.api.domain.work;
 
 import java.time.LocalDate;
-import java.time.Period;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -15,8 +14,6 @@ import com.simleetag.homework.api.domain.work.api.CategoryResources;
 import com.simleetag.homework.api.domain.work.repository.CategoryDslRepository;
 import com.simleetag.homework.api.domain.work.repository.CategoryRepository;
 import com.simleetag.homework.api.domain.work.task.TaskDslRepository;
-import com.simleetag.homework.api.domain.work.task.TaskStatus;
-import com.simleetag.homework.api.domain.work.task.api.TaskRateResponse;
 import com.simleetag.homework.api.domain.work.task.api.TaskResponse;
 import com.simleetag.homework.api.domain.work.taskGroup.TaskGroup;
 import com.simleetag.homework.api.domain.work.taskGroup.TaskGroupRepository;
@@ -91,22 +88,6 @@ public class CategoryService {
 
     public List<TaskResponse> findAllTasksByDueDate(Long memberId, LocalDate date) {
         return TaskResponse.from(taskDslRepository.findAllWithTaskGroupByHomeIdAndOwnerAndDueDate(memberId, date));
-    }
-
-    public List<TaskRateResponse> calculateTaskRatesByDueDates(Long memberId, LocalDate startDate, LocalDate endDate) {
-        Period period = Period.between(startDate, endDate);
-        List<TaskRateResponse> list = new ArrayList<>();
-        for (int i = 0; i < period.getDays() + 1; i++) {
-            double allTasks = taskDslRepository.findAllWithTaskGroupByHomeIdAndOwnerAndDueDate(memberId, startDate.plusDays(i)).size();
-            double doneTasks = taskDslRepository.findAllWithTaskGroupByHomeIdAndOwnerAndDueDate(memberId, startDate.plusDays(i)).stream().filter(task -> task.getTaskStatus().equals(TaskStatus.COMPLETED)).count();
-            double rate = 0;
-            if (allTasks != 0) {
-                rate = doneTasks / allTasks * 100.0;
-            }
-            TaskRateResponse response = new TaskRateResponse(startDate.plusDays(i), (int) rate);
-            list.add(response);
-        }
-        return list;
     }
 
     public CategoryResources.Reply.MeWithTaskGroup deleteCategory(Long categoryId) {
