+++ src/test/java/com/simleetag/homework/api/domain/work/api/CategoryControllerFlow.java
@@ -7,7 +7,6 @@ import java.util.List;
 
 import com.simleetag.homework.api.common.FlowSupport;
 import com.simleetag.homework.api.common.IdentifierHeader;
-import com.simleetag.homework.api.domain.work.task.api.TaskRateResponse;
 import com.simleetag.homework.api.domain.work.task.api.TaskResponse;
 
 import org.springframework.http.MediaType;
@@ -68,24 +67,6 @@ public class CategoryControllerFlow extends FlowSupport {
         return Arrays.asList(objectMapper.readValue(responseBody, TaskResponse[].class));
     }
 
-
-    public List<TaskRateResponse> checkRatesWithDueDate(String invitation, LocalDate startDate, LocalDate endDate, Long memberId) throws Exception {
-        MultiValueMap<String, String> params = new LinkedMultiValueMap<>();
-        params.add("startDate", startDate.toString());
-        params.add("endDate", endDate.toString());
-        params.add("memberId", memberId.toString());
-        final String responseBody = mockMvc.perform(
-                                                   get("/api/categories/tasks/rate")
-                                                           .header(IdentifierHeader.HOME.getKey(), invitation)
-                                                           .params(params)
-                                           ).andExpect(
-                                                   status().isOk()
-                                           )
-                                           .andReturn()
-                                           .getResponse()
-                                           .getContentAsString(StandardCharsets.UTF_8);
-        return Arrays.asList(objectMapper.readValue(responseBody, TaskRateResponse[].class));
-    }
 
     public CategoryResources.Reply.MeWithTaskGroup deleteCategory(String invitation, Long categoryId) throws Exception {
         final String responseBody = mockMvc.perform(
