+++ src/test/java/com/simleetag/homework/api/domain/work/api/CategoryControllerTest.java
@@ -16,11 +16,8 @@ import com.simleetag.homework.api.domain.user.oauth.ProviderType;
 import com.simleetag.homework.api.domain.user.oauth.api.OAuthControllerFlow;
 import com.simleetag.homework.api.domain.user.oauth.api.dto.TokenRequest;
 import com.simleetag.homework.api.domain.user.oauth.api.dto.TokenResponse;
-import com.simleetag.homework.api.domain.work.task.TaskStatus;
 import com.simleetag.homework.api.domain.work.task.api.TaskControllerFlow;
-import com.simleetag.homework.api.domain.work.task.api.TaskRateResponse;
 import com.simleetag.homework.api.domain.work.task.api.TaskResponse;
-import com.simleetag.homework.api.domain.work.task.api.TaskStatusEditRequest;
 import com.simleetag.homework.api.domain.work.taskGroup.Cycle;
 import com.simleetag.homework.api.domain.work.taskGroup.Difficulty;
 import com.simleetag.homework.api.domain.work.taskGroup.TaskGroupType;
@@ -154,71 +151,6 @@ public class CategoryControllerTest extends TestSupport {
             // then
             final int taskSize = categoryController.findAllWithDueDate(home.invitation(), LocalDate.now(), everMemberId).size();
             assertThat(taskSize).isEqualTo(1);
-        }
-
-    }
-
-    @Nested
-    class checkTaskRateTest {
-
-        @Test
-        @DisplayName("TaskStatus - 전체 UNFINISHED일 때 TaskRate 0 확인")
-        void checkTaskRateAllUnfinished() throws Exception {
-
-            //given
-            final List<CategoryResources.Request.Create> request = new ArrayList<>();
-            CategoryResources.Request.Create.CategoryCreateRequest categoryCreateRequest = new CategoryResources.Request.Create.CategoryCreateRequest(null, "새로운 일회성 카테고리");
-            CategoryResources.Request.Create.TaskGroupCreateRequest taskGroupCreateRequest = new CategoryResources.Request.Create.TaskGroupCreateRequest(null, "일회성 집안일", TaskGroupType.TEMPORARY, new Cycle(Collections.singletonList(LocalDate.now().getDayOfWeek()), LocalDate.now(), 0), Difficulty.LOW, 1L, everMemberId);
-            request.add(new CategoryResources.Request.Create(categoryCreateRequest, taskGroupCreateRequest));
-
-            // when
-            categoryController.createNewCategory(home.invitation(), request);
-
-            // then
-            List<TaskRateResponse> taskRateResponse = categoryController.checkRatesWithDueDate(home.invitation(), LocalDate.now(), LocalDate.now(), everMemberId);
-            assertThat(taskRateResponse.get(0).rate()).isEqualTo(0);
-        }
-
-        @Test
-        @DisplayName("TaskStatus - 전체 COMPLETED일 때 TaskRate 100 확인")
-        void checkTaskRateAllCompleted() throws Exception {
-
-            //given
-            final List<CategoryResources.Request.Create> request = new ArrayList<>();
-            CategoryResources.Request.Create.CategoryCreateRequest categoryCreateRequest = new CategoryResources.Request.Create.CategoryCreateRequest(null, "새로운 일회성 카테고리");
-            CategoryResources.Request.Create.TaskGroupCreateRequest taskGroupCreateRequest = new CategoryResources.Request.Create.TaskGroupCreateRequest(null, "일회성 집안일", TaskGroupType.TEMPORARY, new Cycle(Collections.singletonList(LocalDate.now().getDayOfWeek()), LocalDate.now(), 0), Difficulty.LOW, 1L, everMemberId);
-            request.add(new CategoryResources.Request.Create(categoryCreateRequest, taskGroupCreateRequest));
-
-            // when
-            categoryController.createNewCategory(home.invitation(), request);
-            TaskStatusEditRequest taskEditRequest = new TaskStatusEditRequest(TaskStatus.COMPLETED);
-            Long taskId = categoryController.findAllWithDueDate(home.invitation(), LocalDate.now(), everMemberId).get(0).taskId();
-            taskController.changeTaskStatus(home.invitation(), taskId, taskEditRequest);
-
-            // then
-            List<TaskRateResponse> taskRateResponse = categoryController.checkRatesWithDueDate(home.invitation(), LocalDate.now(), LocalDate.now(), everMemberId);
-            assertThat(taskRateResponse.get(0).rate()).isEqualTo(100);
-        }
-
-        @Test
-        @DisplayName("TaskStatus - 전체 중 1/3 COMPLETED일 때 TaskRate 33 확인")
-        void checkTaskRatePartiallyCompleted() throws Exception {
-
-            //given
-            final List<CategoryResources.Request.Create> request = new ArrayList<>();
-            CategoryResources.Request.Create.CategoryCreateRequest categoryCreateRequest = new CategoryResources.Request.Create.CategoryCreateRequest(null, "새로운 일회성 카테고리");
-            CategoryResources.Request.Create.TaskGroupCreateRequest taskGroupCreateRequest = new CategoryResources.Request.Create.TaskGroupCreateRequest(null, "일회성 집안일", TaskGroupType.TEMPORARY, new Cycle(Collections.singletonList(LocalDate.now().getDayOfWeek()), LocalDate.now(), 0), Difficulty.LOW, 1L, everMemberId);
-            request.add(new CategoryResources.Request.Create(categoryCreateRequest, taskGroupCreateRequest, taskGroupCreateRequest, taskGroupCreateRequest));
-
-            // when
-            categoryController.createNewCategory(home.invitation(), request);
-            TaskStatusEditRequest taskStatusEditRequest = new TaskStatusEditRequest(TaskStatus.COMPLETED);
-            Long taskId = categoryController.findAllWithDueDate(home.invitation(), LocalDate.now(), everMemberId).get(0).taskId();
-            taskController.changeTaskStatus(home.invitation(), taskId, taskStatusEditRequest);
-
-            // then
-            List<TaskRateResponse> taskRateResponse = categoryController.checkRatesWithDueDate(home.invitation(), LocalDate.now(), LocalDate.now(), everMemberId);
-            assertThat(taskRateResponse.get(0).rate()).isEqualTo(33);
         }
 
     }
